Resolve CoerceShared field region obligations
