+++ tests/ui/reborrow/coerce-shared-field-region-obligation.stderr
@@ -1,0 +1,18 @@
+error[E0308]: mismatched types
+  --> $DIR/coerce-shared-field-region-obligation.rs:29:1
+   |
+LL | impl<'a> CoerceShared<Target<'a>> for Source<'a> {}
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ lifetime mismatch
+   |
+   = note: expected trait `CoerceShared<FieldRef<'a, &'static ()>>`
+              found trait `CoerceShared<FieldRef<'_, &'a ()>>`
+note: the lifetime `'a` as defined here...
+  --> $DIR/coerce-shared-field-region-obligation.rs:29:6
+   |
+LL | impl<'a> CoerceShared<Target<'a>> for Source<'a> {}
+   |      ^^
+   = note: ...does not necessarily outlive the static lifetime
+
+error: aborting due to 1 previous error
+
+For more information about this error, try `rustc --explain E0308`.
